% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amnfis_classifier.R
\name{fn.amnfis}
\alias{fn.amnfis}
\title{Main function to train a neural network based on the AMNFIS architecture.}
\usage{
fn.amnfis(X, d, clusters)
}
\arguments{
\item{X}{input numeric matrix with the trining data of dimensions mxn where m are the number of examples and n the number of features(attributes)}

\item{d}{input numeric vector with the classes of the training data}

\item{clusters}{matrix with the centroids of each cluster}
}
\value{
object a model which contains the params(weights) of the neural network
}
\description{
Main function to train a neural network based on the AMNFIS architecture.
}
