% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amnfis_classifier.R
\name{fn.amnfis_simulate}
\alias{fn.amnfis_simulate}
\title{Function that makes a prediction of the class given the data}
\usage{
fn.amnfis_simulate(obj, X, C)
}
\arguments{
\item{obj}{input object with the network parameters(weights)}

\item{X}{input numeric matrix which contains the input data}

\item{C}{input numeric matrix with the centroids of each cluster}
}
\value{
y a numeric vector which contains the class predicted for each instance in X
}
\description{
Function that makes a prediction of the class given the data
}
